refactor: break 3 real circular dependencies — 11→2 type-only cycles
Fixed 3 runtime circular requires:
1. ACP session-service ↔ pause-intervention: extracted error classes
   to services/acp/errors.ts
2. ACP index ↔ terminal-bridge: terminal-bridge now imports types
   directly from json-rpc-client.ts and types.ts instead of barrel
3. metrics ↔ metrics-aggregation: extracted SessionForAggregation and
   KeyNameMap to metrics-types.ts, moved computeAggregateMetrics
   consumers to import directly

Remaining 2 cycles (pipeline↔state-store, telegram types↔persistence)
are type-only false positives — verified no circular requires in
compiled JS output.

Before: 11 cycles (7 in session.ts + 4 real)
After: 2 cycles (both type-only, zero runtime impact)

Files changed: 13, 58 insertions, 45 deletions
Tests: 411 files, 5882 passed, 0 failures

Related: #4488, #4489
